requirements.txt
Django>=1.8,<2.0
psycopg2
gunicorn==19.1.1
locustio==0.9.0
pytest-django
